Format 'issue created' email
Include a selection of metadata values in the email.

I needed to pass `config` to `deliver-email` in order to be able to
build a URL to the ticket based on the `endpoint` field.

Also needed to make the field list a vector instead of a list, otherwise
'jzon' interpreted it as being a plist and serialized it to a JSON
object.