Makefile: Remove dependencies from `bundle` target
This seemed to cause the `bundle` target to be run even in the packaged
release tarball. In that case, the 'bundle/' directory should already
exist, and we don't want to rebuild it.

We might also have been able to resolve this by generating the 'bundle/'
tar before the project `git archive` tar, but I couldn't figure out how
to concatenate two tars together with BSD tar.